requirements.txt
undetected-chromedriver
selenium
webdriver-manager
beautifulsoup4
requests
spacy
lxml
APScheduler
newspaper3k
lxml_html_clean